Move bit ops to module.
